requirements.txt
openai-agents
python-dotenv
pydantic
duckduckgo-search
bs4 
requests
rich
